improved error catching
the bot will no longer crash it tries to reply with an error message after being removed from the server..

guilddelete and guildmember remove will no longer throw an error if no profiles exist in the guild or there are no cards in a profile

+++ events/guild/guildDelete.js
@@ -8,14 +8,17 @@ module.exports = {
     async execute(guild) {
         console.log(`Heard ${guild.name} get deleted, or I was removed from it.`)
         const profiles = await profileModel.find({ serverID: guild.id });
-        for (let profile of profiles) {
-            for (let card of profile.cards) {
-                let charData = await charModel.findOne({ name: card.name });
-                charData.numClaims -= 1;
-                await charData.save();
+        if (profiles) {
+            for (let profile of profiles) {
+                if (!profile.cards) break;
+                for (let card of profile.cards) {
+                    let charData = await charModel.findOne({ name: card.name });
+                    charData.numClaims -= 1;
+                    await charData.save();
+                }
             }
+            await profileModel.deleteMany({ serverID: guild.id });
         }
-        await profileModel.deleteMany({ serverID: guild.id });
         await shopModel.deleteOne({ serverID: guild.id });
     },
 };
